requirements.txt
asgiref==3.5.1
certifi==2021.10.8
charset-normalizer==2.0.12
Django==4.0.4
idna==3.3
image==1.5.33
Pillow==9.1.0
pyasn1==0.4.8
pygame==2.1.2
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.10.1
requests==2.27.1
rsa==4.8
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.9
